More quiz refactoring

+++ app_script/Code.js
@@ -1,10 +1,17 @@
+/*
+Version 2.1 (23)
+Date: 5/28/2020
+Authors: FS, AB
+*/
+
 function moveFiles(sourceFileId, folder_name) {
 
   var folders = DriveApp.getFoldersByName(folder_name)
   var folder = null;
   while (folders.hasNext()) {
     folder = folders.next();
-    Logger.log(folder.getName());
+    break;
+    //Logger.log(folder.getName());
   }
 
   if (!folder)
@@ -12,23 +19,33 @@ function moveFiles(sourceFileId, folder_name) {
 
   var targetFolderId = folder.getId();
 
-  Logger.log("***", targetFolderId, folder.getName())
+  //Logger.log("***", targetFolderId, folder.getName())
 
   var file = DriveApp.getFileById(sourceFileId);
   file.getParents().next().removeFile(file);
   DriveApp.getFolderById(targetFolderId).addFile(file);
+  return 1
 }
 
-function createQuiz(name, desc, user, item_list, type) {
+function createQuiz(name, desc, user, item_list, options) {
+   // to keep the format and options not settable by the script
+  sample_forms = [
+    '1HvsLyFZlrAaOFJFhn5u-zcYnIXEXp4A7e5ElBZGSNfI', // correct answer hide
+    '102l6S_GF172zfFx3d2gcWNHF1Bv3N6KcDQfXwQsBAgg' // correct answer show
+  ];
+  type = 0;
+  //Logger.log(options)
+  if (options.show_correct) type=1;  // correct answer show
 
-  sample_forms = ['1Yj55-xMTA9nYnatXb7SxjajrFIw_zCI_SW6TC1o1sXs'];
   var file = DriveApp.getFileById(sample_forms[type]);
   file = file.makeCopy(name)
   var form = FormApp.openById(file.getId());
 
   //var form = FormApp.create(name);
   form.deleteAllResponses()
-  form.deleteItem(0)
+
+  if (!options.name)
+    form.deleteItem(0)  // name
 
 
   form.setTitle(name);
@@ -39,12 +56,13 @@ function createQuiz(name, desc, user, item_list, type) {
   if (desc)
     form.setDescription(desc)
 
+  folder_name = 'User Forms';
   if (user && user.email) {
     form.addEditor(user.email)
     folder_name = user.email + "&" + user.givenName;
   }
 
-  form.setPublishingSummary(true)
+  if (options.summary) form.setPublishingSummary(true)
   total_points = 0;
   for (var i=0; i<item_list.length; i++) {
     var ques = item_list[i];
@@ -60,8 +78,8 @@ function createQuiz(name, desc, user, item_list, type) {
     if (!item) continue;
 
     item.setTitle(ques.question);
-    if (ques.desc) {
-        item.setHelpText(ques.desc);
+    if (ques.desc || ques.description) {
+        item.setHelpText(ques.desc || ques.description);
     }
     if (type == 'MULTIPLE_CHOICE' || type == 'CHECKBOX') {
       var choices = [];
@@ -96,6 +114,7 @@ function createQuiz(name, desc, user, item_list, type) {
     //textItem.setValidation(textValidation);
   }
 
+  if (options.email) form.setCollectEmail(true)
 
   var metadata = {
     id: form.getId(),
@@ -106,6 +125,7 @@ function createQuiz(name, desc, user, item_list, type) {
     // destination_id: form.getDestinationId(),
     editors: form.getEditors(),
     summary_url: form.getSummaryUrl(),
+    folder_name: folder_name
     //updated: file.getLastUpdated().toJSON(),
     //created: file.getDateCreated().toJSON()
   }
@@ -115,11 +135,11 @@ function createQuiz(name, desc, user, item_list, type) {
     metadata: metadata,
     id: form.getId()
   }
-  Logger.log(response)
-  if (folder_name)
-    moveFiles(form.getId(), folder_name)
-  else
-    moveFiles(form.getId(), 'User Forms')
+  Logger.log(metadata)
+  //if (folder_name)
+  //  moveFiles(form.getId(), folder_name)
+  //else
+  //  moveFiles(form.getId(), 'User Forms')
 
   return(response)
 
@@ -173,20 +193,43 @@ function getQuizDetails(form_url, get_responses) {
   }
   //Logger.log(item_list)
   responses = []
+  students = []
   var formResponses = form.getResponses();
   if (get_responses) {
     for (var i = 0; i < formResponses.length; i++) {
       var formResponse = formResponses[i];
       var itemResponses = formResponse.getItemResponses();
-      for (var j = 5; j < itemResponses.length; j++) {
+      for (var j = 0; j < itemResponses.length; j++) {
         var itemResponse = itemResponses[j];
-        Logger.log('Response #%s to the question "%s" was "%s [%s] -> {%s}"',
-            (i + 1).toString(),
-            itemResponse.getItem().getTitle(),
-            itemResponse.getResponse(),
-            itemResponse.getScore(),
-            itemResponse.getItem().asMultipleChoiceItem().getChoices()[0].getValue()
-            );
+        var type = itemResponse.getItem().getType()
+        var response =  itemResponse.getResponse()
+        //Logger.info(typeof(response))
+
+        if (type == 'MULTIPLE_CHOICE' || type == 'CHECKBOX') {
+
+          var resp = {id: parseInt(i + 1), score: 1, answer:response, item_text: itemResponse.getItem().getTitle(), type: type}
+          var q_item = (type == 'MULTIPLE_CHOICE' ? itemResponse.getItem().asMultipleChoiceItem() : itemResponse.getItem().asCheckboxItem());
+          var choices = q_item.getChoices()
+          var points = q_item.getPoints();
+          if (!points || points==0) continue;
+          for (var k = 0; k < choices.length; k++) {
+            var correct_ans = choices[k].isCorrectAnswer()
+            var ch = choices[k].getValue()
+            //Logger.info(ch)
+            //Logger.info(correct_ans)
+            var correct = (typeof(response) === 'string' ? ch == response : response.indexOf(ch) >= 0)
+            //Logger.info(correct)
+            if (correct !== correct_ans)
+               resp['score'] = 0
+          }
+          responses.push(resp)
+          Logger.info(resp)
+        }
+        else {
+          var student = {id: parseInt(i + 1), answer:response, text: itemResponse.getItem().getTitle(), type: type}
+          students.push(student)
+          Logger.info(student)
+        }
       }
     }
   }
@@ -197,7 +240,7 @@ function getQuizDetails(form_url, get_responses) {
     total_points: total_points,
     published_url: form.getPublishedUrl(),
     editor_url: form.getEditUrl(),
-    destination_id: form.getDestinationId(),
+    //destination_id: form.getDestinationId(),
     editors: form.getEditors(),
     summary_url: form.getSummaryUrl(),
     updated: file.getLastUpdated().toJSON(),
@@ -209,15 +252,87 @@ function getQuizDetails(form_url, get_responses) {
     items: item_list,
     responses_count: parseInt(formResponses.length),
     responses: responses,
+    students: students,
     metadata: metadata
   }
-  Logger.log(response)
+  //Logger.log(responses)
+  return(response)
+
+}
+
+
+function getQuizResponses(form_url) {
+  // Open a form by URL.
+  var form = FormApp.openByUrl(form_url);
+
+  responses = []
+  students = []
+  var formResponses = form.getResponses();
+  for (var i = 0; i < formResponses.length; i++) {
+      var formResponse = formResponses[i];
+      var itemResponses = formResponse.getItemResponses();
+      for (var j = 0; j < itemResponses.length; j++) {
+        var itemResponse = itemResponses[j];
+        var type = itemResponse.getItem().getType()
+        var response =  itemResponse.getResponse()
+        //Logger.info(typeof(response))
+
+        if (type == 'MULTIPLE_CHOICE' || type == 'CHECKBOX') {
+
+          var resp = {id: parseInt(i + 1), score: 1, answer:response, item_text: itemResponse.getItem().getTitle(), type: type}
+          var q_item = (type == 'MULTIPLE_CHOICE' ? itemResponse.getItem().asMultipleChoiceItem() : itemResponse.getItem().asCheckboxItem());
+          var choices = q_item.getChoices()
+          var points = q_item.getPoints();
+          if (!points || points==0) continue;
+          for (var k = 0; k < choices.length; k++) {
+            var correct_ans = choices[k].isCorrectAnswer()
+            var ch = choices[k].getValue()
+            //Logger.info(ch)
+            //Logger.info(correct_ans)
+            var correct = (typeof(response) === 'string' ? ch == response : response.indexOf(ch) >= 0)
+            //Logger.info(correct)
+            if (correct !== correct_ans)
+               resp['score'] = 0
+          }
+          responses.push(resp)
+          Logger.info(resp)
+        }
+        else {
+          var student = {id: parseInt(i + 1), answer:response, text: itemResponse.getItem().getTitle(), type: type}
+          students.push(student)
+          Logger.info(student)
+        }
+      }
+  }
+
+  var file = DriveApp.getFileById(form.getId());
+  var metadata = {
+    id: form.getId(),
+    published_url: form.getPublishedUrl(),
+    editor_url: form.getEditUrl(),
+    summary_url: form.getSummaryUrl(),
+    updated: file.getLastUpdated().toJSON(),
+    created: file.getDateCreated().toJSON()
+  }
+  response = {
+    title: form.getTitle(),
+    description: form.getDescription(),
+    responses_count: parseInt(formResponses.length),
+    responses: responses,
+    students: students,
+    metadata: metadata
+  }
+  //Logger.log(responses)
   return(response)
 
 }
 
 function testFunction() {
-   getQuizDetails('https://docs.google.com/forms/d/1eJgoRzOq00jv2D32Jyav7o6cElJ9sdORU_oUrJQj69c/edit', false)
+   getQuizDetails('https://docs.google.com/forms/d/1DEUSZfBvcZIaL4c255z6boYHrNhcbg6A93JQqvUNUzY/edit',true)
+}
+
+function testFunction3() {
+   getQuizResponses('https://docs.google.com/forms/d/1DEUSZfBvcZIaL4c255z6boYHrNhcbg6A93JQqvUNUzY/edit')
 }
 
 function testFunction2() {
@@ -226,6 +341,8 @@ function testFunction2() {
      {'question': "name2?", "options": [{"correct": 0, "choice": "123"}, {"correct": 1,"choice": "1253"}, {"correct": 0,"choice": "123123"},{"correct": 0,"choice": "234234"}], type:'MULTIPLE_CHOICE', required: false},
      {'question': "name3?", "options": [{"correct": 0, "choice": "A"}, {"correct": 1,"choice": "B"}, {"correct": 0,"choice": "C"},{"correct": 0,"choice": "D"}, {"correct": 1,"choice": "E"}], type:'CHECKBOX', feedback_correct: 'Very good!'},
    ]
-   var user = {email: 'farrukh503@gmail.com', givenName: 'Farrukh'};
-   createQuiz("RM Form 5", "Summary of features", user, question_list, 0)
+   var user = null; //{email: 'farrukh503@gmail.com', givenName: 'Farrukh'};
+   var options = {email: 1, name: 0, show_correct: 0, summary: 1 }
+   var results = createQuiz("RM Form 7", "Summary of features", user, question_list, options)
+   Logger.log(results['metadata']['folder_name'])
 }
